tffconnect_backend/requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
djangorestframework==3.14.0
flake8==5.0.4
google-auth==2.14.1
idna==3.4
mccabe==0.7.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.5
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
Pillow==9.3.0
